requirements.txt
yfinance==0.1.59
jsonpath==0.82
seaborn==0.11.1
streamlit_quill==0.0.3
numpy==1.20.2
networkx==2.5.1
wordcloud==1.8.1
python_pptx==0.6.18
feedparser==6.0.2
nltk==3.6.2
matplotlib==3.3.4
requests==2.25.1
plotly==4.14.3
fbprophet==0.7.1
pandas==1.2.4
pytrends==4.7.3
summa==1.2.0
pandas_datareader==0.9.0
psutil==5.8.0
jmespath==0.10.0
streamlit==0.82.0
beautifulsoup4==4.9.3
Pillow==8.2.0
